refactor(afk): merge afk and permafk cogs, fixed permafk not peroperly removing afk, added support for changing afk message without removing afk status
